Create SmallAnimalDB_script.sql
Added the test table to the Database stuff folder so you can create the database and test the gui yourself. Remember to change the connect string inorder to properly connect. I have no advice on what to change it to other than to change DataSource in the connection string. An example of the connection string is in the SQLCode.cs file.